Made the Perl script not using grep
Includes tests for excluded lines and multi lines.

Includes tests that the test runner should fail, these are set as KNOWNBUG
so we can run the regressions folder.

We also don't print out when grep options activated as  no chance of doing
by accident as very clearly different to any other syntax. The printing
messes up the formatting of the output so better to not have it.

Added test script to the Makefile

Also ordered the tests into alphabetical orde.r