Cleaning dependencies of the string solver
`src/util` is a more appropriate place for the `string_exprt` class

We changed the constructor with one type argument to be similar to
other `exprt` constructors.
This constructor was used to generate a struct with
fresh symbols but this is no longer the case, so we use a function
fresh string instead.

Dependencies in the solver that were not needed have been removed.

For some functions this requieres us to now pass the string type as
argument.