[string-refine] Change the concretization process to avoid calling the solver diffblue/test-gen#158
We do this by looking at the indexes for which we have concrete value
because they are in the index set and assigning to the others the
value of the next concrete value.
This is more efficient than calling the solver an extra step and avoid
problems because lemmas gets added.