[string-refine] Defining a limit on the size of counter example strings
At each step of the string refinement we look for counter examples to
the universal axioms in the current model. A problem is that the model
can have very long string. With the limit we add here we ignore
strings of the model that are too long, in order to avoid being stuck
in this phase.
This limit can be smaller than string-max-length and does not impact
the correctness of the algorithm (contrary to string-max-length).