Add an argument to java-throw-runtime-exceptions
This argument is a list of exception classes separated by ':'s.
The instrumentation is then only done for the corresponding classes.
To instrument all runtime exception, one can use 'java.lang.Exception'
or 'java.lang.Throwable' as an argument. This is because all runtime
exceptions inherit from these two classes.