Replace string symbols only for the dependencies
In the string refinement we replace symbols corresponding to arrays of
characters by cannonical element to ensure that symbols resolving to the
same expression are associated the same node.

We should not replace char array in equations that are passed to supert,
since the underlying solver should handle pointers and arrays, and
string_refinementt should not interfer with it.