Class for smt2 asts and functions for creating ASTs
This is to represent the structure of the smt2 instance generated by
smt2_conv.
This will allow separation between the structure and the printing of
these instances.
The structure is not used in this commit but will ultimately be used by
convert_expr.

The class inherits from non_sharing_treet to reuse the structure of
trees. Sharing does not seem to be useful for the usage we will make of
these ASTs.
We don't inherit from irept so that we can get some type safety: in
particular we can't add a irept child to a smt2_astt by mistake.