Extract and document a fix_type function
This makes the code of rename clearer.
The documentation of fix type clarifies what the extracted piece of code
is trying to achieve.
And this piece of code becomes more reusable, in particular if we want
several versions of rename (in particular it would make sense to have
one for each level).