level2t type for results of rename_level2
Makes it clear from the type that rename_level2 rename an expression up
to level2, and allows us to the renaming level in the type.
The argument is now passed by copy and a new expression
is returned, which also makes the interface of the function clearer.
Note that because of that we need a new version of rename_level2
returning a level2t<ssa_exprt>: before, an ssa_exprt was given as
argument and the type left unchanged but new the argument is copied, so
the returned type would only be level2t<exprt>.