rename_level1 takes a copy instead of reference
This makes the interface clearer as there is no argument that is both an
input and an output.
This also corresponds to how the function was used in a lot of
occurences, in which case the code is now simpler.
In the cases where we actually what the transformation to happen
in-place, we can use std::move.

As a side effect we have to define a version that returns a ssa_exprt as
exprt is sometimes not precise enough.