Move symex_decl definition to goto_symex.cpp
Ideally methods of goto_symext should be declared in goto_symex.h and
defined in goto_symex.cpp, otherwise it makes it more difficut to
naviguate the code.
If the goto_symex.cpp files becomes too big then it is a sign that the
class does too many things and should be split into several
classes or functions.